feat: Modify ModelSeparationField to use DecimalField for field_sort_order
- Updated field_sort_order in ModelSeparationField to use DecimalField with max_digits=5 and decimal_places=2
- Replaced IntegerField with DecimalField to provide more granular sorting precision
- Maintained the default value of 1 while allowing for more flexible sorting options

+++ system/models/field.py
@@ -85,7 +85,7 @@ class ModelSeparationField(DbAuditModel, DbUuidModel):
     label_color = models.CharField(verbose_name=_("标题配色"), max_length=32, null=True, blank=True)
     field_auth = models.CharField(verbose_name=_("字段权限"), max_length=128, null=True, blank=True)
     form_grid = models.DecimalField(verbose_name=_("字段宽度"), max_digits=5, decimal_places=2, null=True, blank=True)
-    field_sort_order = models.IntegerField(default=1, verbose_name=_("排序"))
+    field_sort_order = models.DecimalField(default=1, verbose_name=_("排序"), max_digits=5, decimal_places=2)
     field_read_only = models.BooleanField(default=False, verbose_name=_("字段是否只读"))
 
     class Meta:
